Some jsp directories refactoring, added admin's menu
